Potential fix by calling the update every 5 seconds.  Brute force but works.
When the FireFox "suggest history" dropbox is used, neither the blur nor change event is fired.  I have one solution that covers many of the common cases but fails on some others:

* Works - type, down arrow to suggest history, tab
* Works - type, down arrow to suggest history, enter, tab
* Works - type, click
* Works - type, tab
* Fails - down arrow to suggest history, enter, tab
* Fails - down arrow to suggest history, arrow right, tab

I have another solution which is to have a timer update the budget every 5 seconds.  This will make sure that even if the blur/change events are not fired, the values will be updated soon.  It's brute force but works (in FF so far, still need to test cross-browser).  I'm not seeing any performance impact, even with debugging on.

  #1350